refactor: support protocol nonces in 2d pool (#1218)
For supporting extra validations like
https://github.com/tempoxyz/tempo/issues/1082 ideally we need to extend
our custom pool's support to not just 2d transactions but also to
arbitrary AA (and potentially non-AA) transactions.

This PR starts with customizing the pool to allow protocol nonces and
lifting the requirement of non-zero sequence key.

Changes in the PR:
- `nonce_keys` are removed and instead we simply track the
reverse-lookup for 2d nonce slots. instead of relying on transaction
validator to insert the on-chain nonce into `NonceKeys` we now simply
pass it in the `TransactionValidationOutcome::Valid` and take it as
input in `add_transaction`
- 2d pool now processes all state updates, including nonce changes